Criação de teste com JUnit e Mock

+++ ms-pagamentos/src/main/java/br/com/fiap/mspagamentos/service/PagamentoService.java
@@ -4,9 +4,12 @@ import br.com.fiap.mspagamentos.dto.PagamentoDTO;
 import br.com.fiap.mspagamentos.model.Pagamento;
 import br.com.fiap.mspagamentos.model.Status;
 import br.com.fiap.mspagamentos.repository.PagamentoRepository;
+import br.com.fiap.mspagamentos.service.exception.DatabaseException;
 import br.com.fiap.mspagamentos.service.exception.ResourceNotFoundException;
 import org.hibernate.action.internal.EntityActionVetoException;
+import org.hibernate.dialect.Database;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -76,11 +79,47 @@ public class PagamentoService {
 
     @Transactional
     public void delete(Long id) {
-        try {
-        repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Não encontrado o ID: " + id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
         }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Recurso não encontrado");
+        }
+    }
+
+    @Transactional
+    public void confirmPayment(Long id) {
+        // Recupera o pagamento do DB
+        Optional<Pagamento> pagamento = repository.findById(id);
+
+        // valida se existe
+        if (!pagamento.isPresent()) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
+        // seta o status do pagamento como confirmado
+        pagamento.get().setStatus(Status.CONFIRMADO);
+
+        // salva alteração no DB
+        repository.save(pagamento.get());
+    }
+
+    @Transactional
+    public void cancelPayment(Long id) {
+        // Recupera o pagamento no DB
+        Optional<Pagamento> pagamento = repository.findById(id);
+
+        // valida se existe
+        if(!pagamento.isPresent()) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
+
+        // se o status do pagamento
+        pagamento.get().setStatus(Status.CANCELADO);
+
+        // salva  alteracao
+        repository.save(pagamento.get());
     }
 
 }
